+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reverse lookup for pinned relations in dependency task  [#434](https://github.com/p2panda/aquadoggo/pull/434)
 - Persist and maintain index of operation's position in document [#438](https://github.com/p2panda/aquadoggo/pull/438)
 - Introduce `dialer` behaviour with retry logic [#444](https://github.com/p2panda/aquadoggo/pull/444)
+- Introduce peer sampling to the replication service [#463](https://github.com/p2panda/aquadoggo/pull/463)
 
 ### Changed